+++ pkg/state/state.go
@@ -4,9 +4,12 @@ import (
 	"context"
 	"errors"
 	"fmt"
+	"net/url"
+	"strings"
 	"time"
 
 	"github.com/go-logr/logr"
+	"github.com/spf13/afero"
 
 	"github.com/spegel-org/spegel/internal/channel"
 	"github.com/spegel-org/spegel/pkg/metrics"
@@ -32,6 +35,7 @@ func Track(ctx context.Context, ociClient oci.Client, router routing.Router, res
 			return nil
 		case <-tickerCh:
 			log.Info("running scheduled image state update")
+
 			if err := all(ctx, ociClient, router, resolveLatestTag); err != nil {
 				log.Error(err, "received errors when updating all images")
 				continue
@@ -54,8 +58,91 @@ func Track(ctx context.Context, ociClient oci.Client, router routing.Router, res
 	}
 }
 
+func TrackConfiguration(ctx context.Context, fs afero.Fs, configPath string, registryURLs, mirrorURLs []url.URL, resolveTags bool, registriesFilePath string, refreshInterval time.Duration) error {
+	log := logr.FromContextOrDiscard(ctx)
+
+	immediateCh := make(chan time.Time, 1)
+	immediateCh <- time.Now()
+	close(immediateCh)
+	expirationTicker := time.NewTicker(refreshInterval)
+	defer expirationTicker.Stop()
+	tickerCh := channel.Merge(immediateCh, expirationTicker.C)
+	registryURLsMap := map[string]bool{}
+	for _, url := range registryURLs {
+		log.Info("add registry url to containerd configuration", "registry", url.String())
+		registryURLsMap[url.String()] = true
+	}
+	for {
+		select {
+		case <-ctx.Done():
+			return nil
+		case <-tickerCh:
+			if registriesFilePath == "" {
+				log.Info("registries file path not set, skip")
+				continue
+			}
+
+			log.Info("running scheduled configuration update")
+			if err := updateConfiguration(ctx, fs, configPath, registryURLsMap, mirrorURLs, resolveTags, registriesFilePath); err != nil {
+				log.Error(err, "received errors when updating configuration")
+				continue
+			}
+		}
+	}
+}
+
+func updateConfiguration(ctx context.Context, fs afero.Fs, configPath string, registryURLsMap map[string]bool, mirrorURLs []url.URL, resolveTags bool, registriesFilePath string) error {
+	log := logr.FromContextOrDiscard(ctx)
+	if registriesFilePath == "" {
+		log.Info("registries file path not set, skip")
+		return nil
+	}
+	files, err := afero.ReadDir(fs, registriesFilePath)
+	if errors.Is(err, afero.ErrFileNotFound) {
+		log.Info("registries file not found, skip", "path", registriesFilePath)
+		return nil
+	}
+	if err != nil {
+		return err
+	}
+
+	var urlsNeedUpdate []url.URL
+
+	for _, file := range files {
+		// skip hidden files
+		if strings.HasPrefix(file.Name(), "..") {
+			continue
+		}
+		urlRaw := "https://" + file.Name()
+		parsedURL, err := url.Parse(urlRaw)
+		if err != nil {
+			log.Error(err, "failed to parse registry url", "url", urlRaw)
+			continue
+		}
+		err = oci.ValidateRegistry(*parsedURL)
+		if err != nil {
+			log.Error(err, "failed to validate registry url", "url", urlRaw)
+			continue
+		}
+
+		if _, ok := registryURLsMap[urlRaw]; !ok {
+			log.Info("try to add registry url to containerd configuration", "registry", urlRaw)
+			registryURLsMap[urlRaw] = true
+
+			urlsNeedUpdate = append(urlsNeedUpdate, *parsedURL)
+		}
+	}
+
+	err = oci.AddNewMirrorConfiguration(ctx, fs, configPath, urlsNeedUpdate, mirrorURLs, resolveTags)
+	if err != nil {
+		return err
+	}
+
+	return nil
+}
+
 func all(ctx context.Context, ociClient oci.Client, router routing.Router, resolveLatestTag bool) error {
-	log := logr.FromContextOrDiscard(ctx).V(4)
+	// log := logr.FromContextOrDiscard(ctx).V(4)
 	imgs, err := ociClient.ListImages(ctx)
 	if err != nil {
 		return err
@@ -73,7 +160,7 @@ func all(ctx context.Context, ociClient oci.Client, router routing.Router, resol
 		// Handle the list re-sync as update events; this will also prevent the
 		// update function from setting metrics values.
 		event := oci.ImageEvent{Image: img, Type: oci.UpdateEvent}
-		log.Info("sync image event", "image", event.Image.String(), "type", event.Type)
+		// log.Info("sync image event", "image", event.Image.String(), "type", event.Type)
 		keyTotal, err := update(ctx, ociClient, router, event, skipDigests, resolveLatestTag)
 		if err != nil {
 			errs = append(errs, err)