+++ CHANGELOG.md
@@ -12,6 +12,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - [#373](https://github.com/spegel-org/spegel/pull/373) Implement support to disable registry filtering in Containerd.
 - [#717](https://github.com/spegel-org/spegel/pull/717) Extend tests for distribution.
 
+- [#373](https://github.com/spegel-org/spegel/pull/373) Implement support to disable registry filtering in Containerd.
+
 ### Changed
 
 ### Deprecated